Update DrawRouteInWorld function This function is the key function for drawing AR navigation lines. This function is implemented through Unity's LineRenderer component.
Brief description of ideas:
Get the quadrant of the mobile phone's orientation and planned path vector. This quadrant is different from the common Cartesian coordinate system, but is defined in terms of compass angles.
In this algorithm, we need to classify and discuss the two two-dimensional vectors vCam and vNor. When they are in the same quadrant, there is naturally no need to flip the path; when they are not in the same quadrant, further classification discussion is required. When the quadrants of these two two-dimensional vectors are adjacent and the angle does not exceed 60 degrees, we think that there is no need to flip the path at this time; when the quadrants are adjacent and the angle exceeds 60 degrees, we think that flipping is needed at this time. path; and when the quadrant is a diagonal, it is natural to flip the path.